requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-environ==0.11.2
django-filter==24.3
djangorestframework-api-key==3.0.0
drf-spectacular==0.27.2
django-rq==3.0
rq==2.0.0
rq-scheduler==0.13.1
psycopg[binary]==3.2.3
redis==5.2.0
requests==2.32.3
gunicorn==23.0.0
whitenoise==6.8.2
django-prometheus==2.3.1
sentry-sdk[django]==2.18.0
structlog==24.4.0
python-json-logger==2.0.7
